+++ RoboRally/src/utils/json/protocol/CheckPointReachedBody.java
@@ -1,7 +1,14 @@
 package utils.json.protocol;
 
+import com.google.gson.annotations.Expose;
+
+/** This is the wrapper class for the message body of the 'CheckPointReached' protocol JSON message.
+ * @author Manuel Neumayer
+ */
 public class CheckPointReachedBody {
+    @Expose
     private int playerID;
+    @Expose
     private int number;
 
     public CheckPointReachedBody(int playerID, int number) {