+++ RoboRally/src/utils/json/protocol/MovementBody.java
@@ -1,8 +1,16 @@
 package utils.json.protocol;
 
+import com.google.gson.annotations.Expose;
+
+/** This is the wrapper class for the message body of the 'Movement' protocol JSON message.
+ * @author Manuel Neumayer
+ */
 public class MovementBody {
+    @Expose
     private int playerID;
+    @Expose
     private int x;
+    @Expose
     private int y;
 
     public MovementBody(int playerID, int x, int y) {