+++ RoboRally/src/utils/json/protocol/PlayerShootingBody.java
@@ -1,5 +1,8 @@
 package utils.json.protocol;
 
+/** This is the wrapper class for the message body of the 'PlayerSHooting' protocol JSON message.
+ * @author Manuel Neumayer
+ */
 public class PlayerShootingBody {
 
     // Is always empty