+++ RoboRally/src/utils/json/protocol/ShuffleCodingBody.java
@@ -1,6 +1,12 @@
 package utils.json.protocol;
 
+import com.google.gson.annotations.Expose;
+
+/** This is the wrapper class for the message body of the 'ShuffleCoding' protocol JSON message.
+ * @author Manuel Neumayer
+ */
 public class ShuffleCodingBody {
+    @Expose
     private int playerID;
 
     public ShuffleCodingBody(int playerID) {