+++ RoboRally/src/utils/json/protocol/TimerEndedBody.java
@@ -1,8 +1,14 @@
 package utils.json.protocol;
 
+import com.google.gson.annotations.Expose;
+
 import java.util.ArrayList;
 
+/** This is the wrapper class for the message body of the 'TimerEnded' protocol JSON message.
+ * @author Manuel Neumayer
+ */
 public class TimerEndedBody {
+    @Expose
     private ArrayList<Integer> playerIDs;
 
     public TimerEndedBody(ArrayList<Integer> playerIDs) {